test(capture): Move tests to posthog/api
